+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added more descriptive help text to data entry.
+
 ## [0.7.3] - 2021-02-09
 
 ### Changed